requirements-test.txt
pytest>=8.0.0
requests>=2.31.0
prometheus_client>=0.20.0
pyyaml>=6.0
